# api/openapi.yaml
info:
  title: PaySuper payment solution service.
  version: "1.0.0"
  description: Swagger Specification for PaySuper Management API.
servers:
  - url: "https://api.pay.super.com"
    description: Production API
components:
  schemas:
...
    AcceptInviteRequest:
      type: object
      properties:
        token:
          description: "@inject_tag: validate\" require\""
          type: string
        user_id:
          description: "@inject_tag: validate\" required,hexadecimal,len=2\""
          type: string
        email:
          description: "@inject_tag: validate\" required,emai\""
          type: string
    GetMerchantsForUserResponse:
      type: object
      properties:
        status:
          type: integer
        message:
          type: object
          properties:
            code:
              description: "The error code."
              type: string
            message:
              description: "The error short description."
              type: string
            details:
              description: "The error details."
              type: string
        merchants:
          type: array
          items:
            properties:
              id:
                type: string
              name:
                type: string
              role:
                type: string
...
    AcceptInviteResponse:
      type: object
      properties:
        status:
          type: integer
        message:
          type: object
          properties:
            code:
              description: "The error code."
              type: string
            message:
              description: "The error short description."
              type: string
            details:
              description: "The error details."
              type: string
        role:
          type: object
          properties:
            id:
              description: "@inject_tag: json\" i\" validate\"required,hexadecimal,len=2\""
              type: string
            merchant_id:
              description: "@inject_tag: json\" merchant_i\" validate\"required,hexadecimal,len=2\""
              type: string
            role:
              description: "@inject_tag: json\" rol\" validate\"require\""
              type: string
            status:
              description: "@inject_tag: json\" statu\" bson\"statu\" validate\"oneof=invited accepte\""
              type: string
            user_id:
              description: "@inject_tag: json\" user_i\" bson\"user_i\" validate\"required,hexadecimal,len=2\""
              type: string
            first_name:
              description: "@inject_tag: json\" first_nam\" bson\"first_nam\" validate\"require\""
              type: string
            last_name:
              description: "@inject_tag: json\" last_nam\" bson\"last_nam\" validate\"require\""
              type: string
            email:
              description: "@inject_tag: json\" emai\" bson\"emai\" validate\"required,emai\""
              type: string
            created_at:
              type: object
              properties:
                seconds:
                  description: "9999-12-31T23:59:59Z inclusive."
                  type: integer
                nanos:
                  description: "inclusive."
                  type: integer
            updated_at:
              type: object
              properties:
                seconds:
                  description: "9999-12-31T23:59:59Z inclusive."
                  type: integer
                nanos:
                  description: "inclusive."
                  type: integer
...
    ResponseErrorMessage:
      type: object
      properties:
        code:
          description: "The error code."
          type: string
        message:
          description: "The error short description."
          type: string
        details:
          description: "The error details."
          type: string
...
    CheckInviteTokenRequest:
      type: object
      properties:
        token:
          description: "@inject_tag: validate\" require\""
          type: string
        email:
          description: "@inject_tag: validate\" required,emai\""
          type: string
    CheckInviteTokenResponse:
      type: object
      properties:
        status:
          type: integer
        message:
          type: object
          properties:
            code:
              description: "The error code."
              type: string
            message:
              description: "The error short description."
              type: string
            details:
              description: "The error details."
              type: string
        role_id:
          type: string
        role_type:
          type: string
paths:
...
  /auth/api/v1/report_file/download/{file_id}.{file_type}:
    get:
      summary: Export the report file
      operationId: reportFileDownloadPathDownload
      description: "Export the report file into a PDF, CSV, XLSX"
      tags:
      - Report file
      parameters:
      - name: file_id
        description: The unique identifier for the report file.
        in: path
        required: true
        schema:
          type: string
      - name: file_type
        description: "The supported file format (PDF, CSV, XLSX)."
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Returns the report file
          content:
            text/plain:
              schema:
                type: string
        "400":
          description: "Invalid request data (unable to find the file, the file string is incorrect)"
          content:
            application/pdf:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
            text/csv:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
        "401":
          description: Unauthorized request
          content:
            application/pdf:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
            text/csv:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
        "403":
          description: Access denied
          content:
            application/pdf:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
            text/csv:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
        "404":
          description: The file not found
          content:
            application/pdf:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
            text/csv:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
        "500":
          description: Unable to download the file because of the internal server error
          content:
            application/pdf:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
            text/csv:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
...
  /auth/api/v1/user/invite/check:
    post:
      summary: Check the invitation token
      operationId: inviteCheckCheckInvite
      description: Check the invitation token
      tags:
      - User
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckInviteTokenRequest"
      responses:
        "200":
          description: "Returns the user's role ID and type"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckInviteTokenResponse"
        "400":
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
        "401":
          description: Unauthorized request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
  /auth/api/v1/user/invite/approve:
    post:
      summary: Approve the user invitation
      operationId: inviteApproveApproveInvite
      description: Approve the user invitation
      tags:
      - User
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AcceptInviteRequest"
      responses:
        "200":
          description: "Returns the user's role data"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AcceptInviteResponse"
        "400":
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
        "401":
          description: Unauthorized request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
  /auth/api/v1/user/merchants:
    get:
      summary: Get the list of merchants for this user
      operationId: getMerchantsGetMerchants
      description: Get the list of merchants for this user
      tags:
      - User
      responses:
        "200":
          description: Returns the list of merchants
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetMerchantsForUserResponse"
        "400":
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
        "401":
          description: Unauthorized request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
        "403":
          description: Access denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResponseErrorMessage"
...
openapi: "3.0.2"